CCMSPUI-504 | General tidy up, moved createAndSubmitAmendmentForCase to new AmendmentService
